Ver 1.3 / Agent register for all collision paths
- fix registering
- fix finding available time
- register for all collisions in reverse order
- implement time window

+++ scene/IntersectionScene.h
@@ -42,7 +42,7 @@ public:
     void start(int msec = 10);
     void stop();
     void step();
-    qreal getCurrentTime() const { return currentTime_; }
+    int getCurrentTime() const { return currentTime_; }
 
     PathItem* addCarPath(const PainterPath &path,
                             const QPen &pen = QPen(),
